Added `answer` environment
This is just the basic framework needed to typeset answers.  Future
commits will add features to ensure that the questions and answers are
able to be paired correctly.

By default, `answer` environments always reset counters to equations.
This results in warnings for numbering clashes, and a special equation
counter is needed.